refactor: Change order ID from int to uint64_t for ITCH compatibility
- Order::order_id now uint64_t to match NASDAQ ITCH order reference numbers
- Update order_idx map key type accordingly
- Modify all method signatures (cancelOrder, modifyOrder, applyExecution)
- Add applyExecution method for processing ITCH execution messages
- No functional change for existing tests
- Removed constructor from PriceLevel to avoid extra complexity of copying list<Order> orders

+++ include/order.hpp
@@ -3,11 +3,11 @@
 #include <cstdint>
 
 struct Order {
-    int order_id;
+    uint64_t order_id;
     bool is_buy;
     bool is_market; // true = market order
     uint64_t price; // in cents (ignored if is_market)
     int quantity;
 
-    Order(int id, bool is_buy, bool is_market, uint64_t price, int quantity);  // Constructor
+    Order(uint64_t id, bool is_buy, bool is_market, uint64_t price, int quantity);  // Constructor
 };
